requirements.txt
python-telegram-bot==11.1.0
python-dotenv==0.10.3
dialogflow==0.7.2
vk-api==11.6.1